# docs/api/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "CreatePromotionRequest" : {
        "description" : "Promotion creation request",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "category" : {
            "type" : "string"
          },
          "discountAmountPennies" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "discountPercent" : {
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          },
          "discountType" : {
            "enum" : [
              "PERCENTAGE",
              "FLAT_AMOUNT"
            ],
            "type" : "string"
          },
          "label" : {
            "minLength" : 1,
            "type" : "string"
          },
          "shopId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "validFrom" : {
            "format" : "date-time",
            "type" : "string"
          },
          "validUntil" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [
          "shopId",
          "validFrom",
          "validUntil"
        ],
        "type" : "object"
      },
...
      "CreateShopRequest" : {
        "description" : "Shop creation request",
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "bannerUrl" : {
            "type" : "string"
          },
          "deliveryInfo" : {
            "type" : "string"
          },
          "description" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "minimumOrderPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "openingHours" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "phone" : {
            "type" : "string"
          },
          "published" : {
            "type" : "boolean"
          },
          "slug" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "OrderDto" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "customerEmail" : {
            "type" : "string"
          },
          "customerName" : {
            "type" : "string"
          },
          "customerPhone" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "itemCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "notes" : {
            "type" : "string"
          },
          "orderNumber" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "paymentReference" : {
            "type" : "string"
          },
          "paymentStatus" : {
            "enum" : [
              "NONE",
              "PENDING",
              "AUTHORIZED",
              "CAPTURED",
              "FAILED",
              "REFUNDED"
            ],
            "type" : "string"
          },
          "shopId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [
              "DRAFT",
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY",
              "COMPLETED",
              "CANCELLED",
              "REFUNDED"
            ],
            "type" : "string"
          },
          "subtotalPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tenantId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "totalAmountPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "vatAmountPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "vatRate" : {
            "enum" : [
              "ZERO",
              "REDUCED",
              "STANDARD",
              "EXEMPT"
            ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageOrderDto" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderDto"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PagePromotionDto" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/PromotionDto"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "PromotionDto" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "category" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "discountAmountPennies" : {
            "format" : "int32",
            "type" : "integer"
          },
          "discountPercent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "discountType" : {
            "enum" : [
              "PERCENTAGE",
              "FLAT_AMOUNT"
            ],
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "shopId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "validFrom" : {
            "format" : "date-time",
            "type" : "string"
          },
          "validUntil" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PublicPromotionDto" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "discountAmountPennies" : {
            "format" : "int32",
            "type" : "integer"
          },
          "discountPercent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "discountType" : {
            "enum" : [
              "PERCENTAGE",
              "FLAT_AMOUNT"
            ],
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "validUntil" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PublicShopDto" : {
        "properties" : {
          "acceptsCardPayments" : {
            "type" : "boolean"
          },
          "address" : {
            "type" : "string"
          },
          "bannerUrl" : {
            "type" : "string"
          },
          "deliveryFeePennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deliveryInfo" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "freeDeliveryThresholdPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "minimumOrderPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "openingHours" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "phone" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ShopDto" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "bannerUrl" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "deliveryInfo" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "minimumOrderPennies" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "openingHours" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "phone" : {
            "type" : "string"
          },
          "published" : {
            "type" : "boolean"
          },
          "slug" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "string"
          },
          "tenantId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "JWT token from Keycloak. Must contain tenant claim (`tenant_id`, `tenantId`, or `tid`).",
        "scheme" : "bearer",
        "type" : "http"
      },
      "catalog-scopes" : {
        "description" : "Least-privilege machine/integration access via OAuth2 client-credentials scopes (issue #206 [AI-4]). A catalog:read-only token lists products (200) but cannot mutate them (403).",
        "flows" : {
          "clientCredentials" : {
            "scopes" : {
              "catalog:read" : "Read the product catalog (list/get products)",
              "catalog:write" : "Create, update, and delete products and product images",
              "customers:read" : "Reserved for the MCP model — defined-only, not enforced this phase",
              "customers:write" : "Enforced; gates POST /customers (Phase 25 [AI-02])",
              "orders:read" : "Reserved for the MCP model — defined, not yet enforced",
              "orders:write" : "Enforced; gates POST /orders (Phase 25 [AI-02])"
            },
            "tokenUrl" : "http://localhost:8085/realms/jtoye-dev/protocol/openid-connect/token"
          }
        },
        "type" : "oauth2"
      },
      "tenant-header" : {
        "description" : "Dev fallback: UUID of tenant (only used when JWT lacks tenant claim)",
        "in" : "header",
        "name" : "X-Tenant-Id",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "engineering@jtoye.uk",
      "name" : "J'Toye Engineering"
    },
...
    "license" : {
      "name" : "Proprietary",
      "url" : "https://jtoye.uk/license"
    },
    "title" : "J'Toye OaaS Core API",
    "version" : "1.2.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/orders/shop/{shopId}" : {
      "get" : {
        "description" : "Returns paginated orders for a specific shop of the authenticated tenant",
        "operationId" : "getOrdersByShop",
        "parameters" : [
          {
            "in" : "path",
            "name" : "shopId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "pageable",
            "required" : true,
            "schema" : {
              "$ref" : "#/components/schemas/Pageable"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrderDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Get orders by shop",
        "tags" : [
          "Orders"
        ]
      }
    },
...
    "/api/v1/promotions" : {
      "get" : {
        "description" : "Returns a paginated list of promotions for the authenticated tenant. Optional shopId query param narrows to one shop of the tenant, server-side (requires at least STAFF on that shop).",
        "operationId" : "list_2",
        "parameters" : [
          {
            "description" : "Optional shop to narrow the list to",
            "in" : "query",
            "name" : "shopId",
            "required" : false,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagePromotionDto"
                }
              }
            },
            "description" : "Successfully retrieved promotions"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagePromotionDto"
                }
              }
            },
            "description" : "Unauthorized - missing or invalid JWT"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagePromotionDto"
                }
              }
            },
            "description" : "Caller has no grant on the requested shopId"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "List promotions",
        "tags" : [
          "Promotions"
        ]
      },
      "post" : {
        "description" : "Creates a new promotion for the authenticated tenant",
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePromotionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Promotion created successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Validation error"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Create promotion",
        "tags" : [
          "Promotions"
        ]
      }
    },
    "/api/v1/promotions/{id}" : {
      "delete" : {
        "description" : "Deletes a promotion",
        "operationId" : "delete_1",
        "parameters" : [
          {
            "description" : "Promotion ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Promotion deleted successfully"
          },
          "404" : {
            "description" : "Promotion not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Delete promotion",
        "tags" : [
          "Promotions"
        ]
      },
      "get" : {
        "description" : "Returns a single promotion by ID",
        "operationId" : "getById_1",
        "parameters" : [
          {
            "description" : "Promotion ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Promotion found"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Promotion not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Get promotion by ID",
        "tags" : [
          "Promotions"
        ]
      },
      "put" : {
        "description" : "Updates an existing promotion",
        "operationId" : "update_1",
        "parameters" : [
          {
            "description" : "Promotion ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePromotionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Promotion updated successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Validation error"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotionDto"
                }
              }
            },
            "description" : "Promotion not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Update promotion",
        "tags" : [
          "Promotions"
        ]
      }
    },
...
    "/api/v1/public/shops/{slug}" : {
      "get" : {
        "description" : "Get full details of a published shop by its URL slug.",
        "operationId" : "getShop",
        "parameters" : [
          {
            "in" : "path",
            "name" : "slug",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicShopDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get shop details",
        "tags" : [
          "Public Storefront"
        ]
      }
    },
...
    "/api/v1/public/shops/{slug}/promotions" : {
      "get" : {
        "description" : "Returns currently active promotions for a published shop.",
        "operationId" : "getShopPromotions",
        "parameters" : [
          {
            "in" : "path",
            "name" : "slug",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PublicPromotionDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get active promotions",
        "tags" : [
          "Public Storefront"
        ]
      }
    },
...
    "/api/v1/shops/{id}" : {
      "delete" : {
        "description" : "Deletes a shop for the authenticated tenant",
        "operationId" : "delete",
        "parameters" : [
          {
            "description" : "Shop ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Shop deleted successfully"
          },
          "404" : {
            "description" : "Shop not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Delete shop",
        "tags" : [
          "Shops"
        ]
      },
      "get" : {
        "description" : "Returns a single shop by ID for the authenticated tenant",
        "operationId" : "getById",
        "parameters" : [
          {
            "description" : "Shop ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShopDto"
                }
              }
            },
            "description" : "Shop found"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShopDto"
                }
              }
            },
            "description" : "Shop not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Get shop by ID",
        "tags" : [
          "Shops"
        ]
      },
      "put" : {
        "description" : "Updates an existing shop for the authenticated tenant",
        "operationId" : "update",
        "parameters" : [
          {
            "description" : "Shop ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateShopRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShopDto"
                }
              }
            },
            "description" : "Shop updated successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShopDto"
                }
              }
            },
            "description" : "Validation error"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShopDto"
                }
              }
            },
            "description" : "Shop not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Update shop",
        "tags" : [
          "Shops"
        ]
      }
    },
...
    "/public/shops/{slug}" : {
      "get" : {
        "description" : "Get full details of a published shop by its URL slug.",
        "operationId" : "getShop_1",
        "parameters" : [
          {
            "in" : "path",
            "name" : "slug",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicShopDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get shop details",
        "tags" : [
          "Public Storefront"
        ]
      }
    },
...
    "/public/shops/{slug}/promotions" : {
      "get" : {
        "description" : "Returns currently active promotions for a published shop.",
        "operationId" : "getShopPromotions_1",
        "parameters" : [
          {
            "in" : "path",
            "name" : "slug",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PublicPromotionDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get active promotions",
        "tags" : [
          "Public Storefront"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearer-jwt" : [ ]
    },
    {
      "tenant-header" : [ ]
    }
  ],
...
}